moving summarization and propagating to tests
