+++ examples/goldenspike/README.md
@@ -1,3 +1,3 @@
 This directory contains an example notebook showing a simplified version the end-to-end functionality of RAIL.
 
-- [goldenspike.ipynb](https://htmlpreview.github.io/?https://github.com/LSSTDESC/RAIL/blob/master/examples/goldenspike/goldenspike.html) This notebooks chains together the funcitonality of the creation, estimation, evaluation and summarization modules.
+- [goldenspike.ipynb](https://htmlpreview.github.io/?https://github.com/LSSTDESC/RAIL/blob/master/examples/goldenspike/goldenspike.html) This notebooks chains together the funcitonality of the creation, estimation, and evaluation modules.
