+++ rail/__init__.py
@@ -5,5 +5,4 @@ from . import core
 from . import creation
 from . import evaluation
 from . import estimation
-from . import summarization
 from .main import *
