+++ rail/estimation/algos/pointEstimateHist.py
@@ -4,18 +4,19 @@ A summarizer that simple makes a histogram of a point estimate
 
 import numpy as np
 from ceci.config import StageParameter as Param
-from rail.summarization.summarizer import PZtoNZSummarizer
+from ..summarizer import PZtoNZSummarizer
 import qp
 
-class NaiveStack(PZtoNZSummarizer):
+class PointEstimateHist(PZtoNZSummarizer):
     """Summarizer which simply histograms a point estimate
     """
 
-    name = 'NaiveStack'
+    name = 'PointEstimateHist'
     config_options = PZtoNZSummarizer.config_options.copy()
     config_options.update(zmin=Param(float, 0.0, msg="The minimum redshift of the z grid"),
                           zmax=Param(float, 3.0, msg="The maximum redshift of the z grid"),
-                          nzbins=Param(int, 301, msg="The number of gridpoints in the z grid"))
+                          nzbins=Param(int, 301, msg="The number of gridpoints in the z grid"),
+                          point_estimate=Param(str, 'zmode', msg="Which point estimate to use"))
 
     def __init__(self, args, comm=None):
         PZtoNZSummarizer.__init__(self, args, comm=comm)
@@ -24,7 +25,7 @@ class NaiveStack(PZtoNZSummarizer):
     def run(self):
         test_data = self.get_data('input')
         self.zgrid = np.linspace(self.config.zmin, self.config.zmax, self.config.nzbins+1)
-        pdf_vals = test_data.pdf(self.zgrid)
-        yvals = np.expand_dims(np.sum(np.where(np.isfinite(pdf_vals), pdf_vals, 0.), axis=0), 0)
-        qp_d = qp.Ensemble(qp.interp, data=dict(xvals=self.zgrid, yvals=yvals))
+        hist_vals = np.histogram(test_data.ancil[self.config.point_estimate], bins=self.zgrid)[0]
+        qp_d = qp.Ensemble(qp.hist,
+                           data=dict(bins=self.zgrid, pdfs=np.atleast_2d(hist_vals)))
         self.add_data('output', qp_d)
