--- rail/summarization/__init__.py
@@ -1,2 +1,0 @@
-"""Redshift summarization"""
-from .algos import *
