+++ tests/test_summarizers.py
@@ -4,8 +4,8 @@ import copy
 import pytest
 from rail.core.stage import RailStage
 from rail.core.data import QPHandle
-from rail.summarization.algos import naiveStack, pointEstimateHist
-from rail.summarization.algos import varInference
+from rail.estimation.algos import naiveStack, pointEstimateHist
+from rail.estimation.algos import varInference
 
 testdata = 'tests/data/output_BPZ_lite.fits'
 DS = RailStage.data_store
